ValidationDomainClassExtender does not try to extend any static method anymore
While static methods are excluded from the ValidationAdvice and the BeanValidationParameterAdvice this was not the case for the BeanValidationReturnValueAdvice therefore ByteBuddy tried to extend static methods which it can't.
This causes an error "spamming" the console log when the extender is executed in combination with the DomainMirror initialization.
An exception is thrown like this "java.lang.IllegalStateException: Cannot map this reference for static method or constructor start: private static java.lang.Integer io.domainlifecycles.validation.extend.jakarta.VOStaticMethod.calculatePrivate(java.lang.Long)"

This IllegalStateException can't be caught because it is (probably) thrown in another thread.